Bug #45287: phase 2 : 5.0 64 bit compilation warnings
Fixed various compilation warnings when compiling on a 
 64 bit windows.

+++ client/mysql_upgrade.c
@@ -387,7 +387,7 @@ static void find_tool(char *tool_executable_name, const char *tool_name,
       last_fn_libchar -= 6;
     }
 
-    len= last_fn_libchar - self_name;
+    len= (int) (last_fn_libchar - self_name);
 
     my_snprintf(tool_executable_name, FN_REFLEN, "%.*s%c%s",
                 len, self_name, FN_LIBCHAR, tool_name);
